prepare-commit-msg: handle pkgbase and epoch
If these are set, they should be used. Thanks to @anthraxx for noticing
this.

Since they might not be set, source the PKGBUILD and generate this
message in a subshell which does not leak into any further PKGBUILD
parsing. This also fixes a slight bug where the pkgver function from one
PKGBUILD could trick future PKGBUILDs into thinking they are VCS
packages, not upstream releases.