Use better sanity check for 'abort if no commits'
2d4274c (Abort when the repository has no commits., 2018-09-26) tries to
exit gracefully with a helpful error message when it isn't possible to
use `git subtree` because the repository is in the unborn branch state.

This usually happens right after the repository initialization but it
can also happen after `git checkout --orphan=foo`. The current sanity
check detects the former but not the latter.

Use `rev-parse --verify` in order to only proceed if HEAD is resolvable,
detecting both cases.