bash-completion: if there are no pkgnames, don't generate an empty completion
No one really needs to complete " ", but that was exactly what they were
completing. Most noticeable when completing outside of a git repo,
though it happens in a git repo with only files as well. The issue is
that printf, of course, does not care whether there are zero arguments
to interpolate into the formatter, you told it to print something and it
will print it at least once.

See https://github.com/eli-schwartz/aurpublish/pull/5#issuecomment-464125483