Add check for non-default 'aur' alias in ssh config
Since aurpublish no longer uses the "aur:" alias for git commands,
it is necessary to guard against accidental AUR operations for
users who were interacting with repositories other than the
official AUR by setting a hostname in the ssh config.

This patch adds a check for an 'aur' alias in the ssh config that
points anywhere other than the official AUR git server. If that
situation is found, then it will warn the user that the operation
will be made against the AUR git server, and ask the user if they
wish to continue before any git operations are run against the
remote server.